+++ catsoop/__QTYPES__/pythoncode/pythoncode.py
@@ -42,14 +42,20 @@ def get_sandbox(context):
     _execfile(base, context)
 
 
-SCRIPTS = """
+CODEMIRROR_SCRIPTS = """
 <script type="text/javascript" src="BASE/js/codemirror/codemirror.bundle.min.js"></script>
+"""
+
+ACE_SCRIPTS = """
+<script type="text/javascript" src="BASE/js/ace/ace.js"></script>
 """
 
 
 def extra_headers(info):
-    if info["csq_interface"] == "codemirror":
-        return SCRIPTS
+    if info["csq_interface"] == "ace":
+        return ACE_SCRIPTS
+    elif info["csq_interface"] == "codemirror":
+        return CODEMIRROR_SCRIPTS
     else:
         return None
 
@@ -80,7 +86,7 @@ defaults = {
     "csq_soln": 'print("Hello, World!")',
     "csq_tests": [],
     "csq_hint": lambda score, code, info: "",  # post-test hint generator
-    "csq_interface": "codemirror",
+    "csq_interface": "ace",
     "csq_rows": 14,
     "csq_font_size": 16,
     "csq_always_show_tests": False,
@@ -625,7 +631,59 @@ def render_html_codemirror(last_log, **info):
     )
 
 
+def render_html_ace(last_log, **info):
+    name = info["csq_name"]
+    init = last_log.get(name, None)
+    if init is None:
+        init = make_initial_display(info)
+    else:
+        init = get_code(init, info)
+    fontsize = info["csq_font_size"]
+    params = {
+        "name": name,
+        "init": init,
+        "safeinit": init.replace("<", "&lt;"),
+        "height": info["csq_rows"] * (fontsize + 4),
+        "fontsize": fontsize,
+    }
+
+    return (
+        """
+<div class="ace_editor_wrapper" id="container%(name)s">
+<div id="editor%(name)s" name="editor%(name)s" class="embedded_ace_code">%(safeinit)s
+</div></div>
+<input type="hidden" name="%(name)s" id="%(name)s" />
+<input type="hidden" name="%(name)s_log" id="%(name)s_log" />
+<script type="text/javascript">
+    // @license magnet:?xt=urn:btih:0b31508aeb0634b347b8270c7bee4d411b5d4109&dn=agpl-3.0.txt AGPL-v3
+    var log%(name)s = new Array();
+    var editor%(name)s = ace.edit("editor%(name)s");
+    editor%(name)s.setOption("enableKeyboardAccessibility", true);
+    editor%(name)s.setTheme("ace/theme/textmate");
+    editor%(name)s.getSession().setMode("ace/mode/python");
+    editor%(name)s.setShowFoldWidgets(false);
+    editor%(name)s.setValue(%(init)r)
+    document.getElementById("%(name)s").value = editor%(name)s.getValue();
+    editor%(name)s.on("change",function(e){
+        editor%(name)s.getSession().setUseSoftTabs(true);
+        document.getElementById("%(name)s").value = editor%(name)s.getValue();
+    });
+    editor%(name)s.clearSelection()
+    editor%(name)s.getSession().setUseSoftTabs(true);
+    editor%(name)s.on("paste",function(txt){editor%(name)s.getSession().setUseSoftTabs(false);});
+    editor%(name)s.getSession().setTabSize(4);
+    editor%(name)s.setFontSize("%(fontsize)spx");
+    document.getElementById("container%(name)s").style.height = "%(height)spx";
+    document.getElementById("editor%(name)s").style.height = "%(height)spx";
+    editor%(name)s.resize(true);
+    // @license-end
+</script>"""
+        % params
+    )
+
+
 RENDERERS = {
+    "ace": render_html_ace,
     "textarea": render_html_textarea,
     "codemirror": render_html_codemirror,
     "upload": render_html_upload,